Move processing to it's own sub-module
