+++ modules/process/index.js
@@ -1,0 +1,7 @@
+const valid = require('./valid');
+const invalid = require('./invalid');
+
+module.exports = {
+  valid,
+  invalid,
+};
